backend/requirements.txt
# Install all except facenet-pytorch first:
#   pip install -r requirements.txt
# Then install facenet-pytorch bypassing its stale numpy<2 constraint:
#   pip install facenet-pytorch==2.6.0 --no-deps

flask==3.1.3
flask-cors==6.0.2
opencv-python==4.13.0.92
torch==2.11.0
torchvision==0.26.0
numpy==2.4.4
Pillow==12.1.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.48
pgvector==0.4.2
python-dotenv==1.2.2
requests
tqdm
gunicorn
